backend_match/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# backend_match/main.py: 1
fastapi == 0.63.0

# backend_match/main.py: 3
python_socketio == 4.6.1

# Uvicorn server
uvicorn == 0.13.4
psycopg2 == 2.8.6
python-multipart == 0.0.5
